rename index_record.py to record.py
