+++ conda/core/path_actions.py
@@ -36,7 +36,7 @@ from ..gateways.disk.update import backoff_rename, touch
 from ..history import History
 from ..models.channel import Channel
 from ..models.enums import LinkType, NoarchType, PathType
-from ..models.index_record import Link, PackageRecord, PathDataV1, PathsData
+from ..models.records import Link, PackageRecord, PathDataV1, PathsData
 from ..models.match_spec import MatchSpec
 from ..models.package_cache_record import PackageCacheRecord
 from ..models.prefix_record import PrefixRecord
