+++ tests/models/test_index_record.py
@@ -6,7 +6,7 @@ from unittest import TestCase
 from conda.base.context import context
 from conda.common.compat import text_type
 from conda.models.channel import Channel
-from conda.models.index_record import IndexJsonRecord, PackageRecord
+from conda.models.records import IndexJsonRecord, PackageRecord
 from conda.models.prefix_record import PrefixRecord
 
 log = getLogger(__name__)
