+++ tests/models/test_match_spec.py
@@ -13,7 +13,7 @@ from conda.common.compat import on_win
 from conda.exceptions import CondaValueError
 from conda.models.channel import Channel
 from conda.models.dist import Dist
-from conda.models.index_record import IndexRecord, PackageRecord, PackageRef
+from conda.models.records import IndexRecord, PackageRecord, PackageRef
 from conda.models.match_spec import ChannelMatch, MatchSpec, _parse_spec_str
 from conda.models.version import VersionSpec
