+++ tests/models/test_package_info.py
@@ -3,7 +3,7 @@ from unittest import TestCase
 from conda.base.context import context
 from conda.models.channel import Channel
 from conda.models.enums import FileMode, PathType
-from conda.models.index_record import IndexRecord, PathData, PathsData
+from conda.models.records import IndexRecord, PathData, PathsData
 from conda.models.package_info import Noarch, PackageInfo, PackageMetadata
